refactor: rooms の管理を store に移行
